feat: refactor location of Seed trait moved the Seed trait to the options module and updated the usage of it
