+++ src/lib.rs
@@ -23,11 +23,9 @@ pub mod color_dictionary;
 pub mod options;
 
 use color_dictionary::{ColorDictionary, ColorInformation};
-use options::{Color, Luminosity};
+use options::{Color, Luminosity, Seed};
 use rand::rngs::SmallRng;
 use rand::{Rng, SeedableRng};
-use std::collections::hash_map::DefaultHasher;
-use std::hash::{Hash, Hasher};
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct RandomColor {
@@ -286,58 +284,6 @@ impl RandomColor {
             ((s * v / k * 10000.0) / 100.0) as u32,
             (k / 2.0 * 100.0) as u32,
         ]
-    }
-}
-
-pub trait Seed {
-    fn to_value(self) -> u64;
-}
-
-impl Seed for i64 {
-    fn to_value(self) -> u64 {
-        self as u64
-    }
-}
-
-impl Seed for i32 {
-    fn to_value(self) -> u64 {
-        self as u64
-    }
-}
-
-impl Seed for u64 {
-    fn to_value(self) -> u64 {
-        self
-    }
-}
-
-impl Seed for u32 {
-    fn to_value(self) -> u64 {
-        self as u64
-    }
-}
-
-impl Seed for String {
-    fn to_value(self) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        self.hash(&mut hasher);
-        hasher.finish()
-    }
-}
-
-impl Seed for &String {
-    fn to_value(self) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        self.hash(&mut hasher);
-        hasher.finish()
-    }
-}
-
-impl Seed for &str {
-    fn to_value(self) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        self.hash(&mut hasher);
-        hasher.finish()
     }
 }
