+++ src/options/mod.rs
@@ -1,5 +1,7 @@
 mod color;
 mod luminosity;
+mod seed;
 
 pub use self::color::Color;
 pub use self::luminosity::Luminosity;
+pub use self::seed::Seed;
